requirements.txt
requests>=2.28
Pillow>=10.0
folium>=0.15
torch>=2.2
torchvision>=0.17
pandas>=2.2
numpy>=1.26
tqdm>=4.66
geopy>=2.4
